+++ src/elements/db/SendoutElementQuery.php
@@ -148,7 +148,6 @@ class SendoutElementQuery extends ElementQuery
             'campaign_sendouts.replyToEmail',
             'campaign_sendouts.subject',
             'campaign_sendouts.notificationEmailAddress',
-            'campaign_sendouts.googleAnalyticsLinkTracking',
             'campaign_sendouts.mailingListIds',
             'campaign_sendouts.excludedMailingListIds',
             'campaign_sendouts.segmentIds',